refactor: improve code readability by formatting type declarations
In this commit, we've made minor changes to various PHP files in the project in order to improve
code readability. We've standardized the formatting for type declarations (bool, float, int,
string, array) to be separated by a whitespace character on either side of the `|` symbol where
more than one type is possible. No functionality has changed in this commit - this is purely for
ease of reading and maintaining the code base.