feat: runtime currency
In this commit, we have made several changes to improve the handling of currencies within the
application.

- Added runtime currency support to the 'Handler' class in 'src/QUI/ERP/Currency/Handler.php'. This
method provides the ability to get and set the currency during runtime.
- Refactored the 'setUserCurrency' function in 'ajax/setUserCurrency.php'. Removed the use of
namespace on every call to 'getAllowedCurrencies' and made it less complex.
- Added a script to extend the header in 'src/QUI/ERP/Currency/EventHandler.php'. The script
retrieves the default and runtime currencies.
- Updated 'bin/controls/Switch.js'. Commented out the previous code and added a window location
reload call after changing the currency.

These changes ensure that the active user has the correct currency displayed throughout their
session.